广告接口对接，参数类型处理，长整型转化成整型数据

+++ rear/user/src/main/java/com/spaceobj/domain/SysPhoto.java
@@ -21,7 +21,7 @@ import java.io.Serializable;
 public class SysPhoto implements Serializable {
 
   @TableId(value = "photo_id")
-  private Long photoId;
+  private Integer photoId;
 
   @TableField(value = "photo_url")
   private String photoUrl;
